+++ src/CosmoSim/Simulator.py
@@ -31,7 +31,7 @@ from . import CosmoSim
 from .Image import centreImage, drawAxes, crop, annotatePoint, annotateCircle, translateImage
 
 from .Arguments import CosmoParser
-from .Parameters import Parameters
+from . import Parameters
 import pandas as pd
 
 defaultoutcols = [ "index", "filename", "source", "lens", "chi", "R", "phi", "einsteinR", "sigma", "sigma2", "theta", "x", "y" ]
@@ -41,19 +41,22 @@ class GenericSim:
     This is a generic superclass for shared methods.
     """
     def __init__(self,param=None,row=None,name=None,outcols=None,verbose=1):
-        if verbose > 0: print( "[GenericSim] init ..." )
+        if verbose > 1: print( "[GenericSim] init ..." )
         self.verbose = verbose
 
         self.outcols = outcols
 
         self.name = name
         if param is None: 
-            print( "[GenericSim] No parameters give. Using defaults" )
+            if verbose: print( "[GenericSim] No parameters give. Using defaults" )
             param = Parameters()
         self.param = param
         self.directory = param.get( "directory" )
         if self.directory:
             os.makedirs( self.directory, exist_ok=True )
+
+        if self.verbose > 2:
+            print( "[GenericSim] xireference =", self.param.get( "xireference" ) )
 
     def setParameters(self,row):
         """
@@ -67,9 +70,9 @@ class GenericSim:
         Run the simulator with the given data row.
         """
         if not row is None:
-            if self.verbose: print( "[initSim] using row" )
+            if self.verbose > 1: print( "[initSim] using row" )
             self.setParameters( row )
-            if self.verbose: print( "index", row.get( "index", None ) )
+            if self.verbose > 1: print( "index", row.get( "index", None ) )
             self.param.setRow( row )
             try:
                 name = row.name.split(".")[0]
@@ -77,17 +80,18 @@ class GenericSim:
                 name = row["filename"].split(".")[0]
             self.row = row
             self.name = name
-        elif self.verbose: print( "[initSim] row is None" )
+        elif self.verbose > 1: print( "[initSim] row is None" )
         if self.name is None:
             self.name = self.param.get( "name" )
-        if self.verbose: print( "[initSim] item name:", self.name )
+        if self.verbose>1: print( "[initSim] item name:", self.name )
 
         self.runSim()
 
         self.image = self.getDistortedImage( )
         (self.centreimage,self.centrepoint) = centreImage(self.image)
-        print( "[datagen.py] Centre Point", self.centrepoint,
-              "(Centre of Luminence in Planar Co-ordinates)" )
+        if self.verbose: print(
+                "[datagen.py] Centre Point", self.centrepoint,
+                "(Centre of Luminence in Planar Co-ordinates)" )
     def getDistortedImage(self):
         return  self.sim.getDistortedImage( 
                          critical=self.param.get( "criticalcurves" ),
@@ -95,11 +99,11 @@ class GenericSim:
     def setParameters(self):
         raise Exception("Not implemented")
     def runSim(self):
-        if self.verbose: print( "[runSim]" )
+        if self.verbose>2: print( "[runSim]" )
         self.sim.makeSource( self.param )
-        if self.verbose > 0: print ( "[initSim] ready for runSim()\n" ) ;
+        if self.verbose > 1: print ( "[GenericSim] ready for runSim()\n" ) ;
         self.sim.runSim()
-        if self.verbose > 0: print ( "[initSim] runSim() completed\n" ) ;
+        if self.verbose > 1: print ( "[GenericSim] runSim() completed\n" ) ;
     def getActual(self):
         param = self.param
         name = self.name
@@ -114,8 +118,12 @@ class GenericSim:
         cv.imwrite(fn,im)
     def getAnnotated(self,centred=None,cropsize=None):
         """
-        Stub for a future function to get an image with annotations.
+        Get an image with annotations showing key points and the convergence ring.
+        This is incomplete and should be extended with addition annotations and
+        options.
         """
+        if centred is not None:
+            raise NotImplementedError("centred option for getAnnotated() is not implemented yet.")
         im = self.sim.getDistortedImage( critical=True )
         im = annotatePoint( im, self.centrepoint, colour=( 64, 255, 64 ) )
         pt = self.sim.getXiOffset( (0,0) )
@@ -123,18 +131,21 @@ class GenericSim:
         x, y = pt
         convradius = np.sqrt( x*x + y*y )
         im = annotateCircle( im, pt, radius=convradius, colour=( 64, 64, 255 ) )
-        return im
-    def getImage(self,centred=None,cropsize=None):
-        param = self.param
-        if centred is None: centred = param.get( "centred" )
-        if centred:
-            im = self.centreimage
-        else:
-            im = self.image
-        if cropsize is None: cropsize = param.get( "cropsize" )
+        if cropsize is None: cropsize = self.param.get( "cropsize" )
         if cropsize:
-            im = crop(im,int( param.get( "cropsize" ) ))
-        if param.get( "reflines" ):
+            im = crop(im,int( cropsize ) )
+        return im
+    def getImage(self,centred=None,cropsize=None,reflines=None):
+        if centred is None: centred = self.param.get( "centred" )
+        if reflines is None: reflines = self.param.get( "reflines" )
+        if centred:
+            im = self.centreimage.copy()
+        else:
+            im = self.image.copy()
+        if cropsize is None: cropsize = self.param.get( "cropsize" )
+        if cropsize:
+            im = crop(im,int( cropsize ) )
+        if reflines:
             drawAxes(im)
         return im
     def saveImage(self,name=None):