+++ src/simlib/SimulatorModel.cpp
@@ -1,7 +1,9 @@
-/* (C) 2023: Hans Georg Schaathun <georg@schaathun.net> *
+/* (C) 2023,2026: Hans Georg Schaathun <georg@schaathun.net> *
  * Building on code by Simon Ingebrigtsen, Sondre Westbø Remøy,
  * Einar Leite Austnes, and Simon Nedreberg Runde
  */
+
+#define DEBUG 1
 
 #include "cosmosim/Simulator.h"
 #include "simaux.h"
@@ -9,6 +11,7 @@
 #include <thread>
 #include <pybind11/pybind11.h>
 namespace py = pybind11;
+
 
 SimulatorModel::SimulatorModel() :
         CHI(0.5),
@@ -32,9 +35,9 @@ cv::Mat SimulatorModel::getActual() const {
 
 }
 cv::Mat SimulatorModel::getSource() const {
-   std::cout << "[SimulatorModel::getSource()]\n" ;
+   if (DEBUG) std::cout << "[SimulatorModel::getSource()]\n" ;
    if ( NULL == source ) {
-       std::cout << "[SimulatorModel::getSource()] Source not defined.\n" ;
+       if (DEBUG) std::cout << "[SimulatorModel::getSource()] Source not defined.\n" ;
    }
    return source->getImage() ;
 }
@@ -53,7 +56,7 @@ void SimulatorModel::update( ) {
    updateApparentAbs() ;
    if (DEBUG) std::cout << "[SimulatorModel::update] Done updateApparentAbs()\n" ;
    Py_BEGIN_ALLOW_THREADS
-   std::cout << "[SimulatorModel::update] thread section\n" << std::flush ;
+   if (DEBUG) std::cout << "[SimulatorModel::update] thread section\n" << std::flush ;
    updateInner() ;
    Py_END_ALLOW_THREADS
 }
@@ -121,9 +124,9 @@ void SimulatorModel::drawCritical( cv::Mat img ) {
 void SimulatorModel::updateInner( ) {
     cv::Mat imgApparent = getApparent() ;
 
-    std::cout << "[SimulatorModel::updateInner()] R=" << getEtaAbs() 
+    if (DEBUG) std::cout << "[SimulatorModel::updateInner()] R=" << getEtaAbs() 
               << "; CHI=" << CHI << "\n" ;
-    if ( DEBUG ) {
+    if ( DEBUG>1 ) {
       std::cout << "[SimulatorModel::updateInner()] xi=" << getXi()   
               << "; eta=" << getEta() << "; etaOffset=" << etaOffset << "\n" ;
       std::cout << "[SimulatorModel::updateInner()] nu=" << getNu()   
@@ -149,8 +152,9 @@ void SimulatorModel::updateInner( ) {
 /* This just splits the image space in chunks and runs distort() in parallel */
 void SimulatorModel::parallelDistort(const cv::Mat& src, cv::Mat& dst) {
     int n_threads = std::thread::hardware_concurrency() ;
-    if (DEBUG) std::cout << "[SimulatorModel::parallelDistort] " << n_threads << " threads (maskMode="
-                 << maskMode << ")\n" ;
+    if (DEBUG) std::cout 
+       << "[SimulatorModel::parallelDistort] " << n_threads << " threads (maskMode="
+       << maskMode << ")\n" ;
 
     std::vector<std::thread> threads_vec;
     double maskRadius = getMaskRadius() ;
@@ -273,10 +277,10 @@ void SimulatorModel::undistort(const cv::Mat& src, cv::Mat& dst) {
 void SimulatorModel::calculateAlphaBeta() { 
 
     if ( lens == NULL ) {
-        std::cout << "[calculateAlphaBeta] No lens - does nothing.\n" ;
+        if (DEBUG) std::cout << "[calculateAlphaBeta] No lens - does nothing.\n" ;
     } else {
         cv::Point2d xi = getXi() ;
-        std::cout << "[calculateAlphaBeta] [" << xi << "] ... \n" ;
+        if (DEBUG) std::cout << "[calculateAlphaBeta] [" << xi << "] ... \n" ;
         lens->calculateAlphaBeta( xi, nterms ) ;
     }
 }
@@ -286,6 +290,8 @@ void SimulatorModel::calculateAlphaBeta() {
 
 /* A.  Mode setters */
 void SimulatorModel::setMaskMode(bool b) {
+   if (DEBUG) std::cout
+      << "[SimulatorModel::setMaskMode] " << b << std::endl ;
    maskMode = b ; 
 }
 void SimulatorModel::setBGColour(int b) { bgcolour = b ; }
@@ -294,10 +300,12 @@ void SimulatorModel::setBGColour(int b) { bgcolour = b ; }
 void SimulatorModel::setSource(Source *src) {
     // if ( source != NULL ) delete source ;
     // delete segfaults on object created in python
-    if ( NULL == src ) {
-       std::cout << "[SimulatorModel::setSource] NULL source\n" ;
-    } else {
-       std::cout << "[SimulatorModel::setSource] setting source\n" ;
+    if ( DEBUG ) {
+      if ( NULL == src ) {
+         std::cout << "[SimulatorModel::setSource] NULL source\n" ;
+      } else {
+         std::cout << "[SimulatorModel::setSource] setting source\n" ;
+      }
     }
     source = src ;
 }